change tailwind config
