+++ sdk/theme.ts
@@ -1,0 +1,22 @@
+export const theme = {
+    "primary": "#5881CA",
+    "secondary": "#204d74",
+    "accent": "#3A3A3C",
+    "neutral": "#878787",
+    "base-100": "#F2F2F7",
+    "success": "hsl(150 62% 95%)",
+    "warning": "hsl(43 100% 95%)",
+    "error": "hsl(9 100% 95%)",
+    "info": "#1C1C1E",
+  
+    "--rounded-box": "1rem", // border radius rounded-box utility class, used in card and other large boxes
+    "--rounded-btn": "0.5rem", // border radius rounded-btn utility class, used in buttons and similar element
+    "--rounded-badge": "1.9rem", // border radius rounded-badge utility class, used in badges and similar
+    "--animation-btn": "0.25s", // duration of animation when you click on button
+    "--animation-input": "0.2s", // duration of animation for inputs like checkbox, toggle, radio, etc
+    "--btn-text-case": "uppercase", // set default text transform for buttons
+    "--btn-focus-scale": "0.95", // scale transform of button when you focus on it
+    "--border-btn": "1px", // border width of buttons
+    "--tab-border": "1px", // border width of tabs
+    "--tab-radius": "0.5rem", // border radius of tabs
+  };
